fix : fix Integer type to Long type

See merge request s09-webmobile1-sub2/S09P12A206!22

+++ backend/sswm/src/main/java/com/ground/sswm/auth/domain/Auth.java
@@ -16,13 +16,13 @@ public class Auth {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private Integer userId;//unique
+    private Long id;
+    private Long userId;//unique
     private String refreshToken;
     private String accessToken;
 
     @Builder
-    public Auth(Integer userId, String refreshToken, String accessToken) {
+    public Auth(Long userId, String refreshToken, String accessToken) {
         this.userId = userId;
         this.refreshToken = refreshToken;
         this.accessToken = accessToken;
